Unsplit start of file
